Added EndpointClass .hpp, .cpp and made minor changes
Am modificat codul din endpoint.cpp pentru a avea doua fisiere unul de tip hpp si altul cpp, am mai adaugat comentarii la diferite functii si am schimbat tipul temperaturii in float

+++ src/Fridge.cpp
@@ -1,4 +1,4 @@
-#include "Fridge.h"
+#include "Fridge.hpp"
 #include <list>
 #include <time.h>
 
@@ -8,12 +8,12 @@ Fridge::Fridge()
 
 Fridge::~Fridge(){}
 
-int Fridge::getTemp() 
+float Fridge::getTemp() 
 {
     return temp;
 }
 
-void Fridge::setTemp(int t) 
+void Fridge::setTemp(float t) 
 {
     this->temp = t;
 }
